+++ lib/src/features/talker/presentation/widgets/info_bottom_sheet.dart
@@ -156,6 +156,7 @@ class _InfoDescriptionPlaceholder extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) => Column(
+        crossAxisAlignment: CrossAxisAlignment.start,
         children: [
           Text.rich(
             TextSpan(